Oinarrizko transformazioak eginda

+++ 2_fasea/Iturburu-kodea/matrizeak.h
@@ -9,7 +9,7 @@
 GLdouble *mult(GLdouble *m1, GLdouble *m2);
 GLdouble *translazioa(int x, int y, int z);
 GLdouble *biraketa(int x, int y, int z);
-GLdouble *eskalaketa(int x, int y, int z);
+GLdouble *eskalaketa(float x, float y, float z);
 GLdouble *identitatea();
 
 #endif //ITURBURU_KODEA_MATRIZEAK_H
